+++ MacPacker/Model/AppState.swift
@@ -6,7 +6,6 @@
 //
 
 import Foundation
-import KeyboardShortcuts
 import SwiftUI
 
 class AppState: ObservableObject {
@@ -20,22 +19,4 @@ class AppState: ObservableObject {
     
     /// Holds all archives that the app currently takes care of
     var archives: [UUID: Archive2] = [:]
-}
-
-extension KeyboardShortcuts.Name {
-    static let spacePreview = Self("spacePreview", default: .init(.space, modifiers: []))
-}
-
-extension AppState {
-    func setUpEvents() {
-        KeyboardShortcuts.onKeyUp(for: .spacePreview) {
-            print("Space Preview")
-//            if let archive = state.archive,
-//               let selectedItem = state.selectedItems.first,
-//               let url = archive.extractFileToTemp(selectedItem) {
-//                appDelegate.openPreviewerWindow(for: url)
-//            }
-//            return .handled
-        }
-    }
 }